Preserve scale vs fitscale when image size changes
Before, if you are zoomed in and switch images, it would stay at the
same pixel location on the new image unless you were at fitscale.

Now, it sets the scale of the new image to be the same ratio with
respect to fitscale. Thus, as long as the aspect ratio doesn't change,
the view doesn't move.

If the aspect ratio changes, or you switch between vertical and
horizontal, it will map to the nearest area in the union of the two
images' fit coordinates (this is hard to explain but it's simple to
figure out). This is with respect to the top left of the image, not the
center, but I don't feel like screwing with it too much.