Fix images "not loading" in organize
They actually did load, but if you had scrolled down on a day with tons
of images and switched to a day with fewer images, all the images would
be above and you would need to scroll up to see them.

So I just made it return to bounds after anything that could reduce the
number of images in the view.