Fix incorrect histogram updating on slider drag
I first redid the Edit state machine for loading images, cleaning it up
a lot.

That didn't fix it, but then I found that the sliders were updating
position (but not moving) upon mouse release...

That meant that the quick image would be reloaded, but not displayed
because no parameters were changed. Likewise, the full size image would
not be recomputed, so the histogram would be from the low res image and
not the high res one.

Weird, fun, funky. Death to all bugs!