Fix crashes when loading inaccessible images
It now checks for the image being accessible before trying to load the
exif data. In the Filmulate tab, the info and buttons along the top are
replaced with an error message saying which file is not accessible.

There are also a few minor fixes: a little vertical spacing changes for
the info text and background brightness button, due to the new line from
the tab restyling, as well as proper initialization of the file info in
ParameterManager before any image is loaded.