Fix #98: deal with errors to make startup quieter
The binding loops and weird sql errors were caused by spurious calling
of the setters of the processing parameters.

So I just disable the setters until an image is selected.

In addition, I removed a few mistyped declarations of animation easings
and such that were for pulsing brightness; turns out that they were
unnecessary and when I fixed them it didn't look as good. So I just took
them out; the brightness ramps linearly instead of inoutquart like I
intended.