Implement raw histogram (fix issues found by it)
This commit implements a raw histogram visible at the top of the
pipeline. In order to make the blue more visible, I backed the histogram
with a slightly lighter color, which turns out to look great for all the
histograms. So now they're like that too.

By implementing the raw histogram, I finally learned that the white
point given is actually the raw whitepoint before black is subtracted...
so I had to subtract it. This improves highlight recovery.

I also discovered in that investigation that I needed to normalize the
Fuji files to that same whitepoint. Now that I do, they work a *lot*
better.