Fix details of the histograms
The last version lost the clipped ends of the histogram, going to zero
instead.

I also made it so it doesn't obviously drop to zero after clipping: it
goes straight across before dropping down out of view.

I also made it so the line width scales more continuously with uiScale.

Finally I gave it some headroom so that the histograms don't poke out
the top as easily.