Fix bugs when file is not found
When the filename stuff got shuffled around in ParameterManager, the
error text display in Edit (at the top) got screwed up, so the error
never appeared.

Secondly, if you clicked on an errored image when none was loaded, or
immediately after you had begun loading a different image, it would
enter an infinite loop of loading because it thought a slider had moved.

The two "error" states (loading error, and error-because-canceled
processing) are now handled separately.