Fix importing time zone issues
Before I was assuming that libraw... somehow... gave UTC time, or
something, knowing what the camera time zone was. I don't know what I
was thinking.

But now I know that LibRaw and exiv2 give the UTC time *assuming that
the camera's time zone is the computer's time zone*.

So now we correct for that.