Update translations
And fix issue with a particular untranslated string.

Backslashes in the original text break the translation lookup, so it
would always show English. Remove the backslash, and backslashes in the
translated text appear fine...

I just split up the line around the backslash and put the backslash in
as a non-translated string.