Fix broken clustering implementation
When almost all the pixels in a block are very similar (for instance: when they are clipped), then splitting that cluster creates two nearly identical cluster centers. This can cause problems when trying to classify which cluster a pixel belongs to. This led to empty clusters with NaN centers, which created problems further down the line. This code lets us bail out if the cluster centers come too close together. By classifying all but one pixel in a cluster, the 1-pixel cluster gets thrown out. It's possible that there is a bigger issue- why were we trying to split a homogeneous cluster in the first place? But this seems to get things to work