Fix "speckles" in denoised output
There was some numerical instability caused by small values of Wb. Here we fall back to the original pixel values if Wb values are too small. The threshold here is pretty aggressive, so it's possible we'd want to allow slightly lower magnitude values of Wb if it causes problems in the future.