Replace outliers with local mean
In the previous method, outliers had low Wt scores and therefore the proper NLM solution couldn't be found. For those pixels, we used the original values from the A matrix. Now we are using the local patch weighted average instead. This blurs outlier pixels a bit.

Because it was difficult to disentangle, this commit also has some performance improvements for expandDims (making use of compile-time known values) and a split of the omp for clause into omp parallel and omp for. This last change prevents a warning from GCC about putting a cancellation within an implied nowait clause.