Fix monochrome raw handling after WB was revamped
It was getting wacky WB multipliers applied, and it looked like a
cyanotype due to a maxed out blue channel. Kinda neat but not desirable.

NLMeans NR also got fixed for monochrome, because it was applying
nonexistent WB multipliers and getting NaNs.

Finally, useless tools are properly disabled for monochrome images.